+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Breaking change
 
 * In an effort to reduce magical behavior and make `wct` easier to understand, it no longer will automatically serve some resources from its own `npm` dependencies and some resources from the project under test. Instead, all resources are served out of the project under test. This gives the project under test control over its testing dependencies and their versions.
-  * As part of this, wct will also require that the project under test have installations of both the client side bower package and the server side node module. This makes results more reproducible, and ensures that they'll be protected from future breaking changes.
+  * As part of this, wct will also require that the project under test have an installation of the client side web-component-tester bower package. We recommend that all projects also have a dependency on the npm web-component-tester node module, and in a future release will will require it. This is to makes results more reproducible, and ensures that they'll be protected from future breaking changes.
   * This release also unifies the behavior of `wct` and `polyserve`, so if your code works without warnings in one it should work in the other.
   * Calling `replace(...)` will use sinon to stub `document.importNode` until `teardown` is called.